streamlit_app/requirements.txt
streamlit
pandas
Pillow
boto3
